+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 as of 1.2.0.
 
+## [Unreleased]
+
+### Changed
+
+- Making all requests through a single Session/connection pool
+
 ## [1.2.1] - 2022-08-03
 
 ## Changed